move and cleanup
