+++ src/client/src/containers/PrivacyModal/index.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 
 import asModal from "../../components/Modal";
+import { IDependency } from "../DependencyInfo";
 import buttonStyles from "../../css/buttonStyles.module.css";
 import styles from "./styles.module.css";
 import { ReactComponent as Cancel } from "../../assets/cancel.svg";
@@ -13,8 +14,6 @@ import { closeModalAction } from "../../actions/modalActions/modalActions";
 import { AppState } from "../../reducers";
 import { Dispatch } from "redux";
 import RootAction from "../../actions/ActionType";
-
-import { IDependency } from "../DependencyInfo";
 
 interface IStateProps {
   isModalOpen: boolean;
