+++ src/client/src/translations/whitelist_en.json
@@ -11,6 +11,8 @@
   "loginCard.signIn",
   "outputPath.outputPath",
   "outputPath.ariaOutputPath",
+  "selectableCard.details",
+  "selectableCard.selected",
   "summaryTile.changeItemName",
   "topNavBar.newProject",
   "topNavBar.frameworks",
@@ -110,8 +112,6 @@
   "rightSidebar.frontendFramework",
   "rightSidebar.backendFramework",
   "rightSidebar.services",
-  "selectableCard.details",
-  "selectableCard.selected",
   "selectBackendFramework.selectBackendFramework",
   "selectFrontendFramework.selectFrontendFramework",
   "selectPages.pagesTitleQuestion",
